Code splitting and fixes
