+++ src/routers/index.js
@@ -1,0 +1,8 @@
+module.exports = ({ server, config }) => {
+    server.get('/', async () => {
+        return {
+            statusCode: 200,
+            message: `API docs: ${config.docs}`
+        };
+    });
+};
